+++ custom_components/glance_clock/__init__.py
@@ -17,6 +17,7 @@ PLATFORMS = [
     Platform.LIGHT,
     Platform.SELECT,
     Platform.NUMBER,
+    Platform.BUTTON,
 ]